Move AbstractMethod into main package.
