+++ src/main/java/com/nexmo/client/Method.java
@@ -19,10 +19,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.voice.endpoints;
+package com.nexmo.client;
 
-
-import com.nexmo.client.NexmoClientException;
 
 import java.io.IOException;
