+++ src/main/java/com/nexmo/client/sns/SnsEndpoint.java
@@ -21,6 +21,7 @@
  */
 package com.nexmo.client.sns;
 
+import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.NexmoResponseParseException;
@@ -32,7 +33,6 @@ import com.nexmo.client.sns.request.SnsRequest;
 import com.nexmo.client.sns.response.SnsPublishResponse;
 import com.nexmo.client.sns.response.SnsResponse;
 import com.nexmo.client.sns.response.SnsSubscribeResponse;
-import com.nexmo.client.voice.endpoints.AbstractMethod;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
