+++ src/main/java/com/nexmo/client/voice/endpoints/ListCallsMethod.java
@@ -22,6 +22,7 @@
 package com.nexmo.client.voice.endpoints;
 
 
+import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.NexmoUnexpectedException;
