+++ src/main/java/com/nexmo/client/voice/endpoints/StopTalkMethod.java
@@ -21,6 +21,7 @@
  */
 package com.nexmo.client.voice.endpoints;
 
+import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.JWTAuthMethod;
